Fetch proxy deployments
