+++ libs/remix-ui/run-tab/src/lib/constants/index.ts
@@ -46,3 +46,6 @@ export const SET_DEPLOY_OPTIONS = 'SET_DEPLOY_OPTIONS'
 export const SET_CURRENT_CONTRACT = 'SET_CURRENT_CONTRACT'
 export const SET_PROXY_ENV_ADDRESS = 'SET_PROXY_ENV_ADDRESS'
 export const SET_REMIXD_ACTIVATED = 'SET_REMIXD_ACTIVATED'
+export const FETCH_PROXY_DEPLOYMENTS = 'FETCH_PROXY_DEPLOYMENTS'
+export const NEW_PROXY_DEPLOYMENT = 'NEW_PROXY_DEPLOYMENT'
+export const RESET_PROXY_DEPLOYMENTS = 'RESET_PROXY_DEPLOYMENTS'
